requirements.txt
langchain_openai
python-dotenv
langchain_groq
langchain_text_splitters
langchain_community
langchain_chroma
langchainhub
llama-index-embeddings-huggingface
langchain-community
langchain_huggingface